(PC-11643) move breadcrumb to new components
