+++ pro/src/components/pages/Offers/Offer/Thumbnail/ImportTab/ImportTab.jsx
@@ -6,11 +6,11 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 
-import Breadcrumb, { STYLE_TYPE_TAB } from 'components/layout/Breadcrumb'
 import {
   IMPORT_TAB_ID,
   URL_TAB_ID,
 } from 'components/pages/Offers/Offer/Thumbnail/_constants'
+import Breadcrumb, { STYLE_TYPE_TAB } from 'new_components/Breadcrumb'
 
 const ImportTab = ({ activeTab, changeTab, isLoading }) => {
   const steps = [
