+++ pro/src/new_components/Breadcrumb/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Breadcrumb'
+export * from './Breadcrumb'
